refactor(data_access): extract cypher builders
