+++ tests/test_cypher_generation.py
@@ -1,5 +1,5 @@
-from data_access.character_queries import generate_character_node_cypher
-from data_access.world_queries import generate_world_element_node_cypher
+from data_access.cypher_builders.character_cypher import generate_character_node_cypher
+from data_access.cypher_builders.world_cypher import generate_world_element_node_cypher
 from kg_maintainer import CharacterProfile, WorldItem
